# api-docs.yaml
openapi: 3.0.0
info:
  title: Movie API
  version: "1.0.0"
  description: API cho việc lấy danh sách phim trending, tìm kiếm phim và lấy chi tiết phim từ TMDB.
servers:
  - url: http://localhost:3000
tags:
  - name: movies
    description: API về thông tin phim
paths:
  /api/trending:
    get:
      tags:
        - movies
      summary: Lấy danh sách phim trending
      description: >
        Endpoint này trả về danh sách phim trending theo tuần.
        Có thể sử dụng tham số `page` để phân trang.
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Số trang cần lấy dữ liệu (mặc định là 1).
      responses:
        "200":
          description: Danh sách phim trending được trả về thành công.
          content:
            application/json:
              schema:
                type: object
                properties:
                  page:
                    type: integer
                    example: 1
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        backdrop_path:
                          type: string
                          example: "/2Nti3gYAX513wvhp8IiLL6ZDyOm.jpg"
                        id:
                          type: integer
                          example: 950387
                        title:
                          type: string
                          example: "A Minecraft Movie"
                        original_title:
                          type: string
                          example: "A Minecraft Movie"
                        overview:
                          type: string
                          example: "Four misfits find themselves struggling with ordinary problems when they are suddenly pulled through a mysterious portal into the Overworld: a bizarre, cubic wonderland that thrives on imagination. To get back home, they'll have to master this world while embarking on a magical quest with an unexpected, expert crafter, Steve."
                        poster_path:
                          type: string
                          example: "/yFHHfHcUgGAxziP1C3lLt0q2T4s.jpg"
                        media_type:
                          type: string
                          example: "movie"
                        adult:
                          type: boolean
                          example: false
                        original_language:
                          type: string
                          example: "en"
                        genre_ids:
                          type: array
                          items:
                            type: integer
                          example: [10751, 35, 12, 14]
                        popularity:
                          type: number
                          format: float
                          example: 333.9632
                        release_date:
                          type: string
                          format: date
                          example: "2025-03-31"
                        video:
                          type: boolean
                          example: false
                        vote_average:
                          type: number
                          format: float
                          example: 5.9
                        vote_count:
                          type: integer
                          example: 86
                  total_pages:
                    type: integer
                    example: 500
                  total_results:
                    type: integer
                    example: 10000
              examples:
                trendingData:
                  summary: Ví dụ dữ liệu trending
                  value:
                    page: 1
                    results:
                      - backdrop_path: "/2Nti3gYAX513wvhp8IiLL6ZDyOm.jpg"
                        id: 950387
                        title: "A Minecraft Movie"
                        original_title: "A Minecraft Movie"
                        overview: "Four misfits find themselves struggling with ordinary problems when they are suddenly pulled through a mysterious portal into the Overworld: a bizarre, cubic wonderland that thrives on imagination. To get back home, they'll have to master this world while embarking on a magical quest with an unexpected, expert crafter, Steve."
                        poster_path: "/yFHHfHcUgGAxziP1C3lLt0q2T4s.jpg"
                        media_type: "movie"
                        adult: false
                        original_language: "en"
                        genre_ids: [10751, 35, 12, 14]
                        popularity: 333.9632
                        release_date: "2025-03-31"
                        video: false
                        vote_average: 5.9
                        vote_count: 86
                      - backdrop_path: "/wi2CEFSyoxoBWdFeVmTeqnlX3qI.jpg"
                        id: 1233575
                        title: "Black Bag"
                        original_title: "Black Bag"
                        overview: "When intelligence agent Kathryn Woodhouse is suspected of betraying the nation, her husband – also a legendary agent – faces the ultimate test of whether to be loyal to his marriage, or his country."
                        poster_path: "/hHPovtU4b96LHcoeEwRkGHI5btw.jpg"
                        media_type: "movie"
                        adult: false
                        original_language: "en"
                        genre_ids: [18, 9648, 53]
                        popularity: 174.4116
                        release_date: "2025-03-12"
                        video: false
                        vote_average: 6.593
                        vote_count: 204
                    total_pages: 500
                    total_results: 10000
        "500":
          description: Lỗi khi lấy phim trending.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Lỗi khi lấy phim trending
  /api/search:
    get:
      tags:
        - movies
      summary: Tìm kiếm phim
      description: >
        Endpoint này cho phép tìm kiếm phim dựa trên từ khóa được cung cấp.
        Tham số `query` là bắt buộc, `page` là tùy chọn.
      parameters:
        - in: query
          name: query
          required: true
          schema:
            type: string
          description: Từ khóa tìm kiếm phim.
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Số trang cần lấy dữ liệu (mặc định là 1).
      responses:
        "200":
          description: Kết quả tìm kiếm phim được trả về thành công.
          content:
            application/json:
              schema:
                type: object
                properties:
                  page:
                    type: integer
                    example: 1
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        adult:
                          type: boolean
                          example: false
                        backdrop_path:
                          type: string
                          example: "/cyecB7godJ6kNHGONFjUyVN9OX5.jpg"
                        genre_ids:
                          type: array
                          items:
                            type: integer
                          example: [28, 878, 12]
                        id:
                          type: integer
                          example: 1726
                        original_language:
                          type: string
                          example: "en"
                        original_title:
                          type: string
                          example: "Iron Man"
                        overview:
                          type: string
                          example: "After being held captive in an Afghan cave, billionaire engineer Tony Stark creates a unique weaponized suit of armor to fight evil."
                        popularity:
                          type: number
                          format: float
                          example: 19.66
                        poster_path:
                          type: string
                          example: "/78lPtwv72eTNqFW9COBYI0dWDJa.jpg"
                        release_date:
                          type: string
                          format: date
                          example: "2008-04-30"
                        title:
                          type: string
                          example: "Iron Man"
                        video:
                          type: boolean
                          example: false
                        vote_average:
                          type: number
                          format: float
                          example: 7.649
                        vote_count:
                          type: integer
                          example: 26814
                  total_pages:
                    type: integer
                    example: 5
                  total_results:
                    type: integer
                    example: 82
              examples:
                searchData:
                  summary: Ví dụ dữ liệu tìm kiếm phim
                  value:
                    page: 1
                    results:
                      - adult: false
                        backdrop_path: "/cyecB7godJ6kNHGONFjUyVN9OX5.jpg"
                        genre_ids: [28, 878, 12]
                        id: 1726
                        original_language: "en"
                        original_title: "Iron Man"
                        overview: "After being held captive in an Afghan cave, billionaire engineer Tony Stark creates a unique weaponized suit of armor to fight evil."
                        popularity: 19.66
                        poster_path: "/78lPtwv72eTNqFW9COBYI0dWDJa.jpg"
                        release_date: "2008-04-30"
                        title: "Iron Man"
                        video: false
                        vote_average: 7.649
                        vote_count: 26814
                      - adult: false
                        backdrop_path: "/iVped1djsF0tvGkvnHbzsE3ZPTF.jpg"
                        genre_ids: [28, 12, 878]
                        id: 68721
                        original_language: "en"
                        original_title: "Iron Man 3"
                        overview: "When Tony Stark's world is torn apart by a formidable terrorist called the Mandarin, he starts an odyssey of rebuilding and retribution."
                        popularity: 14.3369
                        poster_path: "/qhPtAc1TKbMPqNvcdXSOn9Bn7hZ.jpg"
                        release_date: "2013-04-18"
                        title: "Iron Man 3"
                        video: false
                        vote_average: 6.929
                        vote_count: 22461
                    total_pages: 5
                    total_results: 82
        "400":
          description: Không có từ khóa tìm kiếm.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Không có từ khóa tìm kiếm
        "500":
          description: Không thể lấy dữ liệu phim.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Không thể lấy dữ liệu phim
  /api/movie/{id}:
    get:
      tags:
        - movies
      summary: Lấy chi tiết phim
      description: >
        Endpoint này trả về chi tiết của một bộ phim dựa trên ID được cung cấp.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID của phim cần lấy chi tiết.
      responses:
        "200":
          description: Chi tiết phim được trả về thành công.
          content:
            application/json:
              schema:
                type: object
                properties:
                  adult:
                    type: boolean
                    example: false
                  backdrop_path:
                    type: string
                    example: "/eGX66zonvc4bXg3rM08RUxdYSDx.jpg"
                  belongs_to_collection:
                    type: object
                    nullable: true
                    example: null
                  budget:
                    type: integer
                    example: 0
                  genres:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        name:
                          type: string
                      example:
                        - id: 28
                          name: "Action"
                        - id: 12
                          name: "Adventure"
                        - id: 878
                          name: "Science Fiction"
                  homepage:
                    type: string
                    example: "https://www.superman.com"
                  id:
                    type: integer
                    example: 1061474
                  imdb_id:
                    type: string
                    example: "tt5950044"
                  origin_country:
                    type: array
                    items:
                      type: string
                    example: ["US"]
                  original_language:
                    type: string
                    example: "en"
                  original_title:
                    type: string
                    example: "Superman"
                  overview:
                    type: string
                    example: "Superman, a cub reporter in Metropolis, embarks on a journey to reconcile his Kryptonian heritage with his human upbringing as Clark Kent."
                  popularity:
                    type: number
                    format: float
                    example: 9.7705
                  poster_path:
                    type: string
                    example: "/wPLysNDLffQLOVebZQCbXJEv6E6.jpg"
                  production_companies:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        logo_path:
                          type: string
                          nullable: true
                        name:
                          type: string
                        origin_country:
                          type: string
                      example:
                        - id: 184898
                          logo_path: "/2Z2hiM1ERqFOWRxNxWoJZ8lTxhb.png"
                          name: "DC Studios"
                          origin_country: "US"
                        - id: 94218
                          logo_path: "/zjIyeqPOPmKa8eubquKLiy2y2rR.png"
                          name: "Troll Court Entertainment"
                          origin_country: "US"
                        - id: 11565
                          logo_path: "/S74cHHbIrdmb5M51Pft72Uo9Gz.png"
                          name: "The Safran Company"
                          origin_country: "US"
                  production_countries:
                    type: array
                    items:
                      type: object
                      properties:
                        iso_3166_1:
                          type: string
                        name:
                          type: string
                      example:
                        - iso_3166_1: "US"
                          name: "United States of America"
                  release_date:
                    type: string
                    format: date
                    example: "2025-07-09"
                  revenue:
                    type: integer
                    example: 0
                  runtime:
                    type: integer
                    example: 0
                  spoken_languages:
                    type: array
                    items:
                      type: object
                      properties:
                        english_name:
                          type: string
                        iso_639_1:
                          type: string
                        name:
                          type: string
                      example:
                        - english_name: "English"
                          iso_639_1: "en"
                          name: "English"
                  status:
                    type: string
                    example: "Post Production"
                  tagline:
                    type: string
                    example: "Look up."
                  title:
                    type: string
                    example: "Superman"
                  video:
                    type: boolean
                    example: false
                  vote_average:
                    type: number
                    format: float
                    example: 0
                  vote_count:
                    type: integer
                    example: 0
              examples:
                detailData:
                  summary: Ví dụ dữ liệu chi tiết phim
                  value:
                    adult: false
                    backdrop_path: "/eGX66zonvc4bXg3rM08RUxdYSDx.jpg"
                    belongs_to_collection: null
                    budget: 0
                    genres:
                      - id: 28
                        name: "Action"
                      - id: 12
                        name: "Adventure"
                      - id: 878
                        name: "Science Fiction"
                    homepage: "https://www.superman.com"
                    id: 1061474
                    imdb_id: "tt5950044"
                    origin_country: ["US"]
                    original_language: "en"
                    original_title: "Superman"
                    overview: "Superman, a cub reporter in Metropolis, embarks on a journey to reconcile his Kryptonian heritage with his human upbringing as Clark Kent."
                    popularity: 9.7705
                    poster_path: "/wPLysNDLffQLOVebZQCbXJEv6E6.jpg"
                    production_companies:
                      - id: 184898
                        logo_path: "/2Z2hiM1ERqFOWRxNxWoJZ8lTxhb.png"
                        name: "DC Studios"
                        origin_country: "US"
                      - id: 94218
                        logo_path: "/zjIyeqPOPmKa8eubquKLiy2y2rR.png"
                        name: "Troll Court Entertainment"
                        origin_country: "US"
                      - id: 11565
                        logo_path: "/S74cHHbIrdmb5M51Pft72Uo9Gz.png"
                        name: "The Safran Company"
                        origin_country: "US"
                    production_countries:
                      - iso_3166_1: "US"
                        name: "United States of America"
                    release_date: "2025-07-09"
                    revenue: 0
                    runtime: 0
                    spoken_languages:
                      - english_name: "English"
                        iso_639_1: "en"
                        name: "English"
                    status: "Post Production"
                    tagline: "Look up."
                    title: "Superman"
                    video: false
                    vote_average: 0
                    vote_count: 0
        "500":
          description: Không thể lấy chi tiết phim.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Không thể lấy chi tiết phim
components:
  schemas:
    MovieTrending:
      type: object
      properties:
        backdrop_path:
          type: string
        id:
          type: integer
        title:
          type: string
        original_title:
          type: string
        overview:
          type: string
        poster_path:
          type: string
        media_type:
          type: string
        adult:
          type: boolean
        original_language:
          type: string
        genre_ids:
          type: array
          items:
            type: integer
        popularity:
          type: number
          format: float
        release_date:
          type: string
          format: date
        video:
          type: boolean
        vote_average:
          type: number
          format: float
        vote_count:
          type: integer
    MovieSearch:
      type: object
      properties:
        adult:
          type: boolean
        backdrop_path:
          type: string
        genre_ids:
          type: array
          items:
            type: integer
        id:
          type: integer
        original_language:
          type: string
        original_title:
          type: string
        overview:
          type: string
        popularity:
          type: number
          format: float
        poster_path:
          type: string
        release_date:
          type: string
          format: date
        title:
          type: string
        video:
          type: boolean
        vote_average:
          type: number
          format: float
        vote_count:
          type: integer
    MovieDetail:
      type: object
      properties:
        adult:
          type: boolean
        backdrop_path:
          type: string
        belongs_to_collection:
          type: object
          nullable: true
        budget:
          type: integer
        genres:
          type: array
          items:
            $ref: "#/components/schemas/Genre"
        homepage:
          type: string
        id:
          type: integer
        imdb_id:
          type: string
        origin_country:
          type: array
          items:
            type: string
        original_language:
          type: string
        original_title:
          type: string
        overview:
          type: string
        popularity:
          type: number
          format: float
        poster_path:
          type: string
        production_companies:
          type: array
          items:
            $ref: "#/components/schemas/ProductionCompany"
        production_countries:
          type: array
          items:
            $ref: "#/components/schemas/ProductionCountry"
        release_date:
          type: string
          format: date
        revenue:
          type: integer
        runtime:
          type: integer
        spoken_languages:
          type: array
          items:
            $ref: "#/components/schemas/SpokenLanguage"
        status:
          type: string
        tagline:
          type: string
        title:
          type: string
        video:
          type: boolean
        vote_average:
          type: number
          format: float
        vote_count:
          type: integer
    Genre:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
    ProductionCompany:
      type: object
      properties:
        id:
          type: integer
        logo_path:
          type: string
          nullable: true
        name:
          type: string
        origin_country:
          type: string
    ProductionCountry:
      type: object
      properties:
        iso_3166_1:
          type: string
        name:
          type: string
    SpokenLanguage:
      type: object
      properties:
        english_name:
          type: string
        iso_639_1:
          type: string
        name:
          type: string